x/dex/types/message_limit_buy_test.go
package types

import (
	"testing"

	"github.com/codchen/new-chain-poc/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgLimitBuy_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgLimitBuy
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgLimitBuy{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgLimitBuy{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
